layout files and brightness control added
