+++ app/src/main/res/layout/activity_main.xml
@@ -5,43 +5,42 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
-    android:background="@drawable/front_screen_background"
+    android:background="@drawable/frontscreen_1"
     tools:context=".MainActivity"
     >
 
     <!--    forest items -->
+
+    <!--    ocean items -->
 
     <ImageView
         android:id="@+id/play_personal_theme_icon"
         android:layout_width="40dp"
         android:layout_height="40dp"
         android:layout_alignParentLeft="false"
-        android:layout_marginTop="90dp"
-        android:layout_marginBottom="681dp"
         android:alpha=".75"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toStartOf="@+id/theme1_title"
-        app:layout_constraintHorizontal_bias="0.644"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.175"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0"
+        app:layout_constraintVertical_bias="0.13"
         app:srcCompat="@drawable/custom_icon" />
 
     <TextView
         android:id="@+id/theme1_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginEnd="120dp"
         android:fontFamily="@font/open_sans_semibold"
         android:text="@string/theme1"
         android:textColor="#e0eee0"
         android:textSize="22sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.173"
+        app:layout_constraintStart_toEndOf="@+id/play_personal_theme_icon"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.129" />
-
-    <!--    ocean items -->
 
     <TextView
         android:id="@+id/theme2_title"
@@ -56,7 +55,7 @@
         app:layout_constraintHorizontal_bias="0.593"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.523" />
+        app:layout_constraintVertical_bias="0.483" />
 
     <ImageView
         android:id="@+id/play_forest_theme_icon2"
@@ -66,17 +65,16 @@
         android:alpha=".75"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.518"
+        app:layout_constraintHorizontal_bias="0.329"
         app:layout_constraintStart_toEndOf="@+id/theme2_title"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.528"
+        app:layout_constraintVertical_bias="0.487"
         app:srcCompat="@drawable/forest_icon" />
 
     <TextView
         android:id="@+id/theme3_title"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="24dp"
         android:layout_marginEnd="32dp"
         android:fontFamily="@font/open_sans_semibold"
         android:text="@string/theme3"
@@ -84,25 +82,23 @@
         android:textSize="22sp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.055"
+        app:layout_constraintHorizontal_bias="0.182"
         app:layout_constraintStart_toEndOf="@+id/play_ocean_them_icon3"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.868" />
+        app:layout_constraintVertical_bias="0.852" />
 
     <ImageView
         android:id="@+id/play_ocean_them_icon3"
         android:layout_width="40dp"
         android:layout_height="40dp"
         android:layout_alignParentLeft="false"
-        android:layout_marginEnd="278dp"
-        android:layout_marginBottom="88dp"
         android:alpha=".75"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.849"
+        app:layout_constraintHorizontal_bias="0.175"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0"
+        app:layout_constraintVertical_bias="0.859"
         app:srcCompat="@drawable/ocean" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
